Move some utilities to ktx package
