+++ app/src/main/java/com/walkingforrochester/walkingforrochester/android/ktx/ImageCapture.kt
@@ -1,0 +1,29 @@
+package com.walkingforrochester.walkingforrochester.android.ktx
+
+import android.content.Context
+import androidx.camera.core.ImageCapture
+import androidx.camera.core.ImageCaptureException
+import timber.log.Timber
+import java.io.File
+import kotlin.coroutines.resume
+import kotlin.coroutines.resumeWithException
+import kotlin.coroutines.suspendCoroutine
+
+suspend fun ImageCapture.takePicture(context: Context, fileName: String): File {
+
+    val photoFile = File(context.filesDir, fileName)
+
+    return suspendCoroutine { continuation ->
+        val outputOptions = ImageCapture.OutputFileOptions.Builder(photoFile).build()
+        takePicture(outputOptions, context.executor, object : ImageCapture.OnImageSavedCallback {
+            override fun onImageSaved(output: ImageCapture.OutputFileResults) {
+                continuation.resume(photoFile)
+            }
+
+            override fun onError(ex: ImageCaptureException) {
+                Timber.e(ex, "Image capture failed")
+                continuation.resumeWithException(ex)
+            }
+        })
+    }
+}
